wrgdv3: rename credentials to credentialscontainer
