+++ src/scripts/modules/gooddata-writer-v3/react/components/CredentialsContainer.jsx
